[spi_host, flash] include test header, fix signature mismatch & typos

+++ sw/device/tests/spi_host_flash_test_impl.h
@@ -41,7 +41,7 @@ status_t test_sector_erase(dif_spi_host_t *spi);
  * @param manufacture_id The expected manufacture_id.
  * @return status_t containing either OK or an error.
  */
-status_t test_read_jedec(dif_spi_host_t *spi, uint16_t manufacture_id);
+status_t test_read_jedec(dif_spi_host_t *spi, uint8_t manufacture_id);
 
 /**
  * Send the enable quad mode command.
@@ -86,7 +86,6 @@ status_t test_quad_read(dif_spi_host_t *spi);
 /**
  * Check if the flash supports 4-byte addressing.
  *
- * @param spi A spi host handler.
  * @return True if 4-byte addressing mode is supported.
  */
 bool is_4_bytes_address_mode_supported(void);
